M37: report multi-view tile rects as fractions (fix macOS tile alignment)
The macOS WebView's CSS viewport is not the host content size in points
(measured 1280x559 vs 1280x836 for the same on-screen region), so the
previous pixel-based rects — even with the macOS x1 / Windows xDPR scaling —
placed tiles in the wrong vertical space: the emerald audio-focus ring sat
well below its video in both Grid and Focus.

Fix: the frontend now reports each tile rect as a fraction (0..1) of the
player area, and the compositor resolves it against its own live drawable
size each frame (Rect::to_px). This maps the DOM overlay and the composited
video onto the same screen region regardless of any viewport-vs-host size or
scale difference — and is mathematically identical to the old behavior on
Windows (fraction x client_size == px x DPR), so it can't regress Windows.
Single playback (rect = fill) is unaffected.

- models.rs / types: TileRect x/y/w/h are now f32 fractions.
- compositor: Rect holds fractions; Rect::to_px resolves against (cw, ch);
  render loop + add_tile use it. Drawable-fill (None) path unchanged.
- mv_add_tile drops its placeholder rect args (it was always overwritten by
  mv_set_rects the moment the tile mounts); the backend uses a zero
  placeholder until the grid reports the real fraction.
- MultiView sends fractions (no more DPR/isWindows scaling).

Removes the temporary diagnostics. Builds clean (cargo + npm). Pending
owner visual confirmation that tiles/ring now align on macOS.

+++ src-tauri/src/models.rs
@@ -377,16 +377,19 @@ pub struct TileState {
     pub state: MpvState,
 }
 
-/// A tile's destination rectangle in the window, reported by the frontend grid
-/// (CSS client coordinates: top-left origin, +y down).
+/// A tile's destination rectangle reported by the frontend grid as **fractions
+/// (0..1)** of the player area (top-left origin, +y down). Fractions, not pixels,
+/// so the compositor maps them onto its own live drawable size — robust to the
+/// WebView viewport and the host surface differing in size/scale (e.g. macOS,
+/// where the CSS viewport is not the content size in points). See `mpv::compositor`.
 #[derive(Debug, Clone, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct TileRect {
     pub tile_id: u32,
-    pub x: i32,
-    pub y: i32,
-    pub w: i32,
-    pub h: i32,
+    pub x: f32,
+    pub y: f32,
+    pub w: f32,
+    pub h: f32,
 }
 
 /// App settings (spec §15 settings keys), returned as a single object so the
